14:25 15.03.2022

+++ PersonProject/Entities/Contact.cs
@@ -6,7 +6,7 @@ namespace PersonProject.Entities
     {
         public int Id { get; set; }
         public int PersonId { get; set; }
-        public decimal PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public string EmailAddress { get; set; }
         public string Location { get; set; }
         public string Information { get; set; }
